+++ src/index.html
@@ -28,6 +28,6 @@
         <div id="modalContent"></div>
       </div>
     </div>
-    <footer>Created by Issac and Luis under</footer>
+    <footer>Created by Issac and Luis</footer>
   </body>
 </html>
