+++ src/modules/navListeners.js
@@ -1,13 +1,14 @@
 import storage from './storage.js';
+import update from './update.js';
 
 export default () => {
   document.getElementById('seasonOne').addEventListener('click', () => {
     storage.getSeasonOne();
-    storage.getLikes();
+    update.getLikes();
   });
 
   document.getElementById('seasonTwo').addEventListener('click', () => {
     storage.getSeasonTwo();
-    storage.getLikes();
+    update.getLikes();
   });
 };
